requirements.txt
docopt
numpy
dpu-utils>=0.1.30
tensorflow-gpu>=1.13.1